refactor: consolidate everything to single file
